services/collector/pyproject.toml
[project]
name = "design-generator-collector"
version = "0.1.0"
description = "High-quality design collection: Playwright crawlers, HuggingFace datasets, auto-ingest"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "playwright>=1.50",
    "httpx>=0.28",
    "Pillow>=11",
    "datasets>=3.3",
    "huggingface-hub>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "rich>=13",
    "typer>=0.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
